model: use floats to store node/leaf bounding boxes

+++ include/model.h
@@ -141,7 +141,7 @@ typedef struct mnode_s {
     int contents;		// 0, to differentiate from leafs
     int visframe;		// node needs to be traversed if current
 
-    short minmaxs[6];		// for bounding box culling
+    float minmaxs[6];		// for bounding box culling
 
     struct mnode_s *parent;
 
@@ -160,7 +160,7 @@ typedef struct mleaf_s {
     int contents;		// wil be a negative contents number
     int visframe;		// node needs to be traversed if current
 
-    short minmaxs[6];		// for bounding box culling
+    float minmaxs[6];		// for bounding box culling
 
     struct mnode_s *parent;
